add pending&notified state to db entries to deal with notify failing

+++ ldf_adapter/approval/__init__.py
@@ -2,13 +2,23 @@
 Implement approval workflow for user deployments.
 """
 import logging
-from typing import List, Optional
+from typing import List, Optional, Tuple
+from enum import Enum
 
 from .. import backend
 from ..userinfo import UserInfo
-from .db import PendingUser, PendingGroup, PendingMemberships, PendingDB
+from .db import PendingUser, PendingGroup, PendingMemberships, SqlitePendingDB, DeploymentState
+
 
 logger = logging.getLogger(__name__)
+
+
+class NotificationType(Enum):
+    NOOP = 0
+    NEW = 1
+    UPDATE_REQUEST = 2
+    UPDATE_GROUPS = 3
+    UNKNOWN = 4
 
 
 class PendingDeployment:
@@ -19,28 +29,28 @@ class PendingDeployment:
     _groups: List[PendingGroup] = []
     _memberships: Optional[PendingMemberships] = None
 
-    def __init__(self, pending_db: PendingDB, userinfo: UserInfo) -> None:
+    def __init__(self, userinfo: UserInfo, user_db_location: str) -> None:
         """Initialise a pending deployment for a federated user.
         If a request already exists for this user, initialise properties.
 
         Args:
-            pending_db (PendingDB): database containing pending deployment requests
             userinfo (UserInfo): user info of federated user
+            user_db_location (str): path to pending db
         """
+        self._pending_db = SqlitePendingDB(location=user_db_location)
         self.unique_id = userinfo.unique_id
         self._sub = userinfo.sub
         self._iss = userinfo.iss
         self._email = userinfo.email
         self._full_name = userinfo.full_name
 
-        self._pending_db = pending_db
-        self._user = pending_db.get_user(userinfo.unique_id)
-        self._memberships = pending_db.get_memberships(userinfo.unique_id)
+        self._user = self._pending_db.get_user(userinfo.unique_id)
+        self._memberships = self._pending_db.get_memberships(userinfo.unique_id)
         if self._memberships:
             self._groups = list(
                 filter(
                     None,
-                    [pending_db.get_group(m) for m in self._memberships.supplementary_groups],
+                    [self._pending_db.get_group(m) for m in self._memberships.supplementary_groups],
                 )
             )
         else:
@@ -103,7 +113,11 @@ class PendingDeployment:
     def name_taken(self, name: str) -> bool:
         """Whether the username is already reserved by *another* user."""
         entry = self._pending_db.get_user_by_username(name)
-        if entry and entry.unique_id != self.unique_id and entry.state == "pending":
+        if (
+            entry
+            and entry.unique_id != self.unique_id
+            and entry.state in [DeploymentState.PENDING, DeploymentState.NOTIFIED]
+        ):
             return True
         return False
 
@@ -118,7 +132,7 @@ class PendingDeployment:
             email=self.email,
             full_name=self.full_name,
             username=service_user.name,
-            state="pending",
+            state=DeploymentState.PENDING,
             cmd=service_user.create_tostring(),
             infodict={},
         )
@@ -140,7 +154,7 @@ class PendingDeployment:
         if group.name in [g.name for g in self.groups]:
             return False
         pending_group = PendingGroup(
-            name=group.name, state="pending", cmd=group.create_tostring(), infodict={}
+            name=group.name, state=DeploymentState.PENDING, cmd=group.create_tostring(), infodict={}
         )
         self._pending_db.add_group(pending_group)
         self._groups.append(pending_group)
@@ -157,14 +171,16 @@ class PendingDeployment:
 
         Return True if the pending db has been modified and False otherwise.
         """
+        if supplementary_groups == [] and removal_groups == []:
+            return False
         supplementary_group_names = [g.name for g in supplementary_groups]
         removal_group_names = [g.name for g in removal_groups]
 
-        membership = PendingMemberships(
+        memberships = PendingMemberships(
             unique_id=service_user.unique_id,
             supplementary_groups=supplementary_group_names,
             removal_groups=removal_group_names,
-            state="pending",
+            state=DeploymentState.PENDING,
             cmd=service_user.mod_tostring(
                 supplementary_groups=supplementary_groups, removal_groups=removal_groups
             ),
@@ -172,18 +188,18 @@ class PendingDeployment:
         )
 
         if self._memberships is None:
-            self._pending_db.add_memberships(membership)
-            self._memberships = membership
+            self._pending_db.add_memberships(memberships)
+            self._memberships = memberships
             return True
         elif set(self._memberships.supplementary_groups) != set(supplementary_group_names) or set(
             self._memberships.removal_groups
         ) != set(removal_group_names):
-            self._pending_db.update_memberships(membership)
-            self._memberships = membership
+            self._pending_db.update_memberships(memberships)
+            self._memberships = memberships
             return True
         return False
 
-    def remove_pending_data(self):
+    def remove_data(self):
         """Remove from pending db all data associated to this user."""
         if self.user:
             self._pending_db.remove_user(self.user.unique_id)
@@ -196,23 +212,23 @@ class PendingDeployment:
 
     def is_pending(self) -> bool:
         """Whether the deployment was requested and is pending approval."""
-        entry = self._pending_db.get_user(self.unique_id)
-        if entry:
-            return entry.state == "pending"
-        return False
+        return self.user is not None and self.user.state in [
+            DeploymentState.PENDING,
+            DeploymentState.NOTIFIED,
+        ]
 
     def is_rejected(self) -> bool:
         """Whether the deployment was requested and was rejected."""
-        entry = self._pending_db.get_user(self.unique_id)
-        if entry:
-            return entry.state == "rejected"
-        return False
+        return self.user is not None and self.user.state == DeploymentState.REJECTED
 
-    def groups_pending(self) -> bool:
-        """Whether a group creation and membership change is pending approval."""
-        return self.groups != [] or self.memberships != None
+    def mod_pending(self) -> bool:
+        """Whether a group membership change is pending approval."""
+        return self.memberships is not None and self.memberships.state in [
+            DeploymentState.PENDING,
+            DeploymentState.NOTIFIED,
+        ]
 
-    def accept(self):
+    def accept(self) -> None:
         """Accept this pending deployment and create user and groups (backend-specific)."""
         for group in self.groups:
             backend.Group.create_fromstring(group.cmd)  # type: ignore
@@ -220,9 +236,9 @@ class PendingDeployment:
             backend.User.create_fromstring(self.user.cmd)  # type: ignore
         if self.memberships:
             backend.User.mod_fromstring(self.memberships.cmd)  # type: ignore
-        self.remove_pending_data()
+        self.remove_data()
 
-    def reject(self):
+    def reject(self) -> None:
         """Reject this pending deployment by setting the state of the user in the pending db
         to 'rejected', and remove pending group and membership entries for this user."""
         if self.user:
@@ -230,6 +246,43 @@ class PendingDeployment:
             self._user = self._pending_db.get_user(self.unique_id)
         for group in self.groups:
             self._pending_db.remove_group(group.name)
-        self._pending_db.remove_memberships(self.unique_id)
         self._groups = []
-        self._memberships = None
+        if self.memberships:
+            self._pending_db.remove_memberships(self.unique_id)
+            self._memberships = None
+
+    def get_notification_type(self) -> Tuple[NotificationType, str]:
+        """Return the type of notification to be sent to the user."""
+        if self.user is None and self.memberships is None:
+            # no pending user or group membership, nothing to notify
+            return NotificationType.NOOP, "No pending deployment, nothing to notify."
+        if self.user and self.user.state == DeploymentState.PENDING:
+            # new deployment request
+            return NotificationType.NEW, "Request for deployment was submitted for approval."
+        if self.memberships and self.memberships.state == DeploymentState.PENDING:
+            # only membership changes are pending
+            if self.user is None:
+                # user exists, but groups need to be updated
+                return (
+                    NotificationType.UPDATE_GROUPS,
+                    "Request for updating user groups was submitted for approval.",
+                )
+            if self.user and self.user.state == DeploymentState.NOTIFIED:
+                # deployment request already notified, but groups have changed and need to be updated
+                return (
+                    NotificationType.UPDATE_REQUEST,
+                    "Updated request for deployment was submitted for approval.",
+                )
+        # nothing to do, request already notified and no changes happened
+        return NotificationType.NOOP, "Request for deployment was already submitted for approval."
+
+    def set_notified(self) -> None:
+        """Set the state of this pending deployment to 'notified' by setting the state of the user
+        and/or its memberships in the pending db to 'notified'.
+        """
+        if self.user:
+            self._pending_db.notify_user(self.unique_id)
+        if self.memberships:
+            self._pending_db.notify_memberships(self.unique_id)
+        for group in self.groups:
+            self._pending_db.notify_group(group.name)
